Update for A-MoreOOP - Add Parser, Storage and Ui classes
