+++ src/main/java/TaskList.java
@@ -1,10 +1,4 @@
-import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.FileNotFoundException;
-import java.util.Scanner;
 
 public class TaskList {
     private ArrayList<Task> list;
@@ -27,65 +21,6 @@ public class TaskList {
 
     public void deleteTaskAt(int index) {
         this.list.remove(index);
-    }
-
-    public void readFromFile() {
-        try {
-            File f = new File("data/duke.txt");
-            Scanner s = new Scanner(f);
-
-            this.list = new ArrayList<>();
-
-            while (s.hasNext()) {
-                String line = s.nextLine();
-                if (line == null || line.equals("")) {
-                    break;
-                }
-                String[] params = line.split(" \\| ");
-                boolean isDone = params[1].equals("1");
-                switch (params[0]) {
-                case "T":
-                    this.addToList(new Todo(params[2], isDone));
-                    break;
-
-                case "D":
-                    this.addToList(new Deadline(params[2], isDone, LocalDateTime.parse(params[3], Duke.TIME_FORMAT)));
-                    break;
-
-                case "E":
-                    this.addToList(new Event(params[2], isDone, LocalDateTime.parse(params[3], Duke.TIME_FORMAT),
-                            LocalDateTime.parse(params[4], Duke.TIME_FORMAT)));
-                    break;
-
-                default:
-                    isDone = params[0].equals("1");
-                    this.addToList(new Task(params[1].trim(), isDone));
-                    break;
-                }
-            }
-
-            s.close();
-        } catch (FileNotFoundException e) {
-            System.out.println("\n" + "OOPS!!! Could not find the file duke.txt");
-        }
-    }
-
-    public void writeToFile() {
-        try {
-            FileWriter fw = new FileWriter("data/duke.txt");
-
-            StringBuffer write = new StringBuffer("");
-            for (Task t: this.list) {
-                write.append(t.getTextRepresentation() + "\n");
-            }
-            fw.write(write.toString());
-
-            fw.close();
-        } catch (FileNotFoundException e) {
-            System.out.println("\n" + "OOPS!!! Could not find the file duke.txt");
-        } catch (IOException e) {
-            System.out.println("\n" + "OOPS!!! " + e.getMessage());
-        }
     }
 
     @Override
